+++ tests/Support/Car.php
@@ -8,8 +8,10 @@ final class Car
 {
     public ?ColorInterface $color = null;
 
-    public function __construct(private EngineInterface $engine, private array $moreEngines = [])
-    {
+    public function __construct(
+        private EngineInterface $engine,
+        private array $moreEngines = []
+    ) {
     }
 
     public function getEngine(): EngineInterface
