+++ tests/Support/EngineMarkOne.php
@@ -8,8 +8,9 @@ final class EngineMarkOne implements EngineInterface
 {
     public const NAME = 'Mark One';
 
-    public function __construct(private int $number = 0)
-    {
+    public function __construct(
+        private int $number = 0
+    ) {
     }
 
     public function getName(): string
