+++ tests/Support/GearBox.php
@@ -9,7 +9,8 @@ namespace Yiisoft\Definitions\Tests\Support;
  */
 final class GearBox
 {
-    public function __construct(private int $maxGear = 5)
-    {
+    public function __construct(
+        private int $maxGear = 5
+    ) {
     }
 }
