+++ tests/Support/NullableInterfaceDependency.php
@@ -6,8 +6,9 @@ namespace Yiisoft\Definitions\Tests\Support;
 
 final class NullableInterfaceDependency
 {
-    public function __construct(private ?EngineInterface $engine)
-    {
+    public function __construct(
+        private ?EngineInterface $engine
+    ) {
     }
 
     public function getEngine(): ?EngineInterface
