+++ tests/Support/UnionCar.php
@@ -6,8 +6,9 @@ namespace Yiisoft\Definitions\Tests\Support;
 
 final class UnionCar
 {
-    public function __construct(private NonExistingEngine|EngineMarkOne|EngineMarkTwo $engine)
-    {
+    public function __construct(
+        private NonExistingEngine|EngineMarkOne|EngineMarkTwo $engine
+    ) {
     }
 
     public function getEngine(): NonExistingEngine|EngineMarkOne|EngineMarkTwo
